move team stuff to challenge folder
